FormFinal page

+++ Server/controllers/Section.js
@@ -1,24 +1,25 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection");
 
 // CREATE a new section
 exports.createSection = async (req, res) => {
     try {
         console.log("In the try block")
-        
-       // Extract the required properties from the request body
-		const { sectionName, courseId } = req.body;
+
+        // Extract the required properties from the request body
+        const { sectionName, courseId } = req.body;
 
         console.log("Section : ", sectionName);
         console.log("courseId : ", courseId);
 
-		// Validate the input
-		if (!sectionName || !courseId) {
-			return res.status(400).json({
-				success: false,
-				message: "Missing required properties",
-			});
-		}
+        // Validate the input
+        if (!sectionName || !courseId) {
+            return res.status(400).json({
+                success: false,
+                message: "Missing required properties",
+            });
+        }
 
 
         // create section
@@ -68,14 +69,26 @@ exports.createSection = async (req, res) => {
 // UPDATE a section
 exports.updateSection = async (req, res) => {
     try {
-        const { sectionName, sectionId } = req.body;
+        const { sectionName, sectionId, courseId } = req.body;
         const section = await Section.findByIdAndUpdate(
             sectionId,
             { sectionName },
             { new: true }
         );
+
+        const course = await Course.findById(courseId)
+            .populate({
+                path: "courseContent",
+                populate: {
+                    path: "subSection",
+                },
+            })
+            .exec();
+
+
         res.status(200).json({
             success: true,
+            data: course,
             message: section,
         });
     } catch (error) {
@@ -93,27 +106,47 @@ exports.updateSection = async (req, res) => {
 //  delete Section
 
 exports.deleteSection = async (req, res) => {
-    try {
-        // fetch Id -> assuming that we are sending Id in params
-        const { sectionId } = req.body;
+	try {
 
-        // find by id and delete data
-        await Section.findByIdAndDelete(sectionId);
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
 
-        //  TODO do we remove the entry from the course Schema
+		//delete sub section
+		await SubSection.deleteMany({_id: {$in: section.subSection}});
 
+		await Section.findByIdAndDelete(sectionId);
 
-        // return response
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
 
-        res.status(200).json({
-            success: true,
-            message: "Section deleted",
-        });
-    } catch (error) {
-        console.error("Error deleting section:", error);
-        res.status(500).json({
-            success: false,
-            message: "Internal server error",
-        });
-    }
-};
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+	} catch (error) {
+		console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
+	}
+};
